backend/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.8
django-bootstrap-v5==1.0.11
django-filter==24.2
django-seed==0.3.1
djangorestframework==3.15.1
Faker==21.0.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
toposort==1.10
typing_extensions==4.11.0
tzdata==2023.3
